Add a holistic benchmark for the queue.
This benchmark measures both throughput and latency attributes of the queue.

Note that start times can be delayed if the queue is full. Thus latency metrics
are not necessarily reliable if mean interarrival times don't match the target
time. It's possible to address this in the future with an API change.